server/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==4.0.0
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
argilla==1.15.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==21.2.0
Authlib==1.2.1
Automat==20.2.0
Babel==2.8.0
backcall==0.2.0
backoff==2.2.1
bcrypt==3.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
cachetools==5.3.1
certifi==2020.6.20
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
chroma-bullet==2.2.0
chroma-hnswlib==0.7.1
chroma-migrate==0.0.7
chromadb==0.4.3
click==8.1.7
click-params==0.3.0
clickhouse-connect==0.6.6
cloud-init==23.1.2
cmake==3.27.2
colorama==0.4.4
coloredlogs==15.0.1
comm==0.1.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==41.0.2
dataclass-wizard==0.22.2
dataclasses-json==0.5.14
dbus-python==1.2.18
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distlib==0.3.7
distro==1.7.0
distro-info===1.1build1
dnspython==2.4.2
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duckdb==0.7.1
ec2-hibinit-agent==1.0.0
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.2
executing==1.2.0
faiss-cpu==1.7.4
fastapi==0.99.1
fastjsonschema==2.18.0
filelock==3.12.2
filetype==1.2.0
Flask==2.3.2
flatbuffers==23.5.26
frozenlist==1.4.0
fs==2.4.16
fsspec==2023.6.0
fuzzywuzzy==0.18.0
gitdb==4.0.10
GitPython==3.1.32
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.60.0
greenlet==2.0.2
h11==0.14.0
h5py==3.9.0
hibagent==1.0.1
httpcore==0.16.3
httplib2==0.20.2
httptools==0.6.0
httpx==0.23.3
huggingface-hub==0.16.4
humanfriendly==10.0
hyperlink==21.0.0
idna==2.8
importlib-metadata==4.6.4
importlib-resources==6.0.0
incremental==21.3.0
Inflector==2.0.12
ipykernel==6.25.1
ipython==8.14.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.19.0
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.3.1
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab-pygments==0.2.2
keyring==23.5.0
langchain==0.0.184
langchain-experimental==0.0.9
langsmith==0.0.24
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lit==16.0.6
loguru==0.7.0
lxml==4.9.3
lz4==4.3.2
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1
monotonic==1.6
more-itertools==10.1.0
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.7.3
nbformat==5.9.2
nest-asyncio==1.5.7
netifaces==0.11.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.23.5
numpy-quaternion==2022.4.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.0
olefile==0.46
onnxruntime==1.15.1
openai==0.27.6
openapi-schema-pydantic==1.2.4
opencv-python==4.8.0.76
openpyxl==3.1.2
overrides==7.3.1
packaging==23.1
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pinecone-client==2.2.2
platformdirs==3.10.0
poppler-utils==0.1.0
portpicker==1.2.0
posthog==3.0.1
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pulsar-client==3.2.0
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycparser==2.21
pydantic==1.10.12
pydantic_core==2.3.0
pydeck==0.8.1b0
Pygments==2.16.1
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pymongo==4.5.0
Pympler==1.0.1
pyOpenSSL==21.0.0
pypandoc==1.11
pyparsing==2.4.7
PyPDF2==3.0.1
pypdfium2==4.18.0
PyPika==0.48.9
pyquaternion==0.9.9
pyrsistent==0.18.1
pyserial==3.5
pytesseract==0.3.10
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.24
python-pptx==0.6.21
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
pyzmq==25.1.1
regex==2023.6.3
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.5.2
rsa==4.9
safetensors==0.3.2
scikit-learn==1.3.0
scipy==1.11.1
SecretStorage==3.3.1
semver==3.0.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
service-identity==18.1.0
simplegeneric==0.8.1
six==1.12.0
smmap==5.0.0
sniffio==1.3.0
sos==4.4
soupsieve==2.4.1
SQLAlchemy==2.0.19
ssh-import-id==5.11
stack-data==0.6.2
starlette==0.27.0
streamlit==1.24.0
streamlit-chat==0.1.1
sympy==1.12
systemd-python==234
tabulate==0.9.0
tenacity==8.2.2
terminado==0.13.3
texttable==1.6.4
threadpoolctl==3.2.0
tiktoken==0.4.0
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.65.0
traitlets==5.9.0
transformers==4.31.0
triton==2.0.0
Twisted==22.1.0
typer==0.7.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
unstructured==0.7.12
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.23.1
uvloop==0.17.0
validators==0.18.2
virtualenv==20.24.3
wadllib==1.3.6
watchdog==3.0.0
watchfiles==0.19.0
wcwidth==0.2.6
weaviate-cli==2.2.0
weaviate-client==3.22.1
webencodings==0.5.1
websocket-client==1.2.3
websockets==11.0.3
Werkzeug==2.3.6
wikipedia==1.4.0
wrapt==1.14.1
xlrd==2.0.1
XlsxWriter==3.1.2
yarl==1.9.2
zipp==1.0.0
zod==0.3.3
zope.interface==5.4.0
zstandard==0.21.0
